pyproject.toml
[project]
name = "objectdetection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.6.1",
    "pillow>=12.1.0",
    "pycocotools>=2.0.11",
    "pydantic>=2.12.5",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
]
